+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - A shiny new splashscreen now tries to mask the unfathomably long loading times
 - KeePass import now creates entries for binary attachments
 - **KeePass export** is now available in the folder bars menu
+- The settings page now remembers the list of recent repositories
 
 #### Changes
 - The *My Favorites* link and the tool icons now remain visible when scrolling